[Polyam] Rename CSS classes for profile redesign
Port df64716b3407cc423cd261126350ed238206febd to polyam-glitch
