+++ app/javascript/flavours/polyam/features/account_timeline/components/account_name.tsx
@@ -21,7 +21,7 @@ import { useRelationship } from '@/flavours/polyam/hooks/useRelationship';
 import { useAppDispatch, useAppSelector } from '@/flavours/polyam/store';
 import FollowerIcon from '@/images/icons/icon_follower.svg?react';
 
-import classes from './redesign.module.scss';
+import classes from './styles.module.scss';
 
 const messages = defineMessages({
   lockedInfo: {
